Disable HTTP Keepalive in API availability tests
HTTP keepalive was actually resulting in TCP connections staying open
for longer than they should have. Investigation revealed that when
redeploying the routers the connections were held open for longer than
the two minute drain time allowed by GoRouter. This was resulting in
occassional failures.

We believe this TCP behaviour is not representative of normal use, so it
is justifyable to remove HTTP keepalive.