Deploy paas-admin application
We want to deploy the paas-admin application for the tenants.

We consider that the admin tool is part of every platform, so we
decided to add the deployment as part of the paas-cf deployment
pipeline.

For the time being we deploy the latest copy of `master` from
the paas-admin branch, but we will add tags as soon as
the paas-admin project has a consistent and automated testing
and tagging pipeline.

We use spruce to render the final manifest with the right
credentials, which is simpler and easier to read than embedded ruby
or similar.

We use the autopilot[1] cf plugin for zero downtime deployment,
as autopilot would create and use the routes defined in the
manifest.

The application is deployed in the route `admin.<system_domain>`.

We will implement logic to somehow retrieve the notify secrets
in a follow up commit.