Add test for apps that take a long time to stage
When initially trying to move Cloud Controller behind GoRouter[1] we had test
failures due to the short idle timeout on the ELBs. The tests only happened to
have caught the failure due to a `cf push` that took a long time.

After a bit of investigation it seems the CF CLI monitors staging via polling,
so it is likely the observed failures were due to something else, like an
upload that timed out for example. However, this test has been left because no
tests exist for this yet and it seems useful.

[1] https://www.pivotaltracker.com/n/projects/1275640/stories/155856825