Remove the use of ENABLE_BILLING_APP flag
We came to the conclusion, there is no real reason to restrict deploying
this application to all of our environments. It takes more time to
deploy that application when it is actually needed than if it would be
done for us every morning.

We also thought, that leaving ability to disable deployment of the
package, clearly indicates problems with our deployment and should be
targeted when it becomes an issue instead.