Use smaller service plan for non-prod enviroments
Now that the billing app will be deployed to all environments at all
times, we could save some money by restricting the service plan it
should use for different environments.

We went with the smallest available plan "Free" for dev and staging,
which should be updated once we rename the naming as we don't think it
should require more resources for small environments.