Do not import AWS creds from an expired STS token
We have a script `./scripts/create_sts_token.sh` for generating
temporary AWS credentials. It acts as a 2-factor way of logging into
an environment, using permanent AWS credentials and a one-time code.

When creating an STS token we calculate when it will expire. If that
script is re-run it will only create a new STS token if the existing
one has expired. If they have expired it will delete the old file and
then try to create a new one.

However if your shell imports the token file on startup you have a
problem. If your temporary credentials have expired then the script
will delete the file but fail to generate new credentials (because the
STS file has overriden your global credentials.) The way around this
is to then open a second shell after that file has been deleted, but
it's all a mess.

In this commit we change the STS Token files to only import the
environment variables if they have not yet expired. The new behaviour
would be that the renewal script works immediately. Many of us use
wrapper scripts but I still encounter related issues sometimes.