Add secrets and create DB for networking services
The cf-networking services require a backing DB for the
network_connectivity and network_policy services.

We will use the same RDS instance than the api, bbs and UAA,
we generate secrets for the new schemas, and create them in the
same way.

In the upstream cf-deployment they use Bosh variables, but
we will keep using our secret generation script for the tiem
being to avoid refactoring now how the DBs are created.