Add cf-networking manifest config
We copy the same configuration as upstream in cf-deployment[1]
for the v1.21 on which our deployment is based.

We copy the same certificates as Bosh variables, but use the DB
secrets generated with our own script.

We also create the required clients for the policy.

We use the RDS postgres instance created in terraform. The DBs
were created in our own external script.

[1] https://github.com/cloudfoundry/cf-deployment/blob/v1.21.0/cf-deployment.yml

Generate client for the network-policy

Required for the network policy service. The secret is used
internally in the manifest, so we can use the Bosh variable
interpolation.