Use initial version and initial content for S3 resources instead of s3init.sh
The default S3 resourse doesn't handle an initial version and content and we
had to initialise these files in separate task using s3init.sh.

We already had a fork of the S3 resource to be able to use IAM instance
profiles for authentication [1], so we added new parameters to the resource so
it can emit an initial version with initial content if no versions exist.

We can pass binary content using the initial_content_binary parameter and
passing the base64 encoded value. We do this for tar.gz files where we pass an
empty tar.gz.

[1] https://github.com/alphagov/paas-s3-resource