Fix `get-cf-cli-config` task's CC API url
In [1] our cf-deployment work removed the `cc.srv_api_uri` manifest
property. This property was still in use by one of the Concourse tasks
and this task no longer exported a valid URL for the CC API. This led
to `destroy-cloudfoundry` not deleting two RDS instances, and thus
failing when trying to destroy the VPC.

This commit fixes the task using a pattern copied from the
`delete_admin.yml` task. There is some deduplication possible here
but this was intended as a small task.

[1] https://github.com/alphagov/paas-cf/commit/bc65bb092e2cd859d57720e30f6b5d34c1d54e5e