Test that every CA cert variable has also a _old
We want to be sure that we append a <name>_old certificate to
support CA rotation. In this commit we verify that every CA
has a <name>_old defined, and it complements the other test
that verifies that both certs are appended whenever they are used.

We need to add a _old certificate for the application_ca certificate
although that certificate does not seem to be used.