Add cf-deployment as submodule
We are going to start tracking the upstream cf-deployment
to generate our manifest.

The easiest approach is keep it as a submodule, as it will
work in concourse, travis and locally.

We point to the version that we are tracking currently, v1.21.0

We also ignore manifests/cf-deployment from shellcheck, as it
fails the checks but it is a third party module.