Add opsfile to rename all the variables of secrets
We are using a external script to generate the secrets, as there
was not the bosh interpolate feature back then.

In cf-deployment these variables are called differently, but used
in the same places. In this commit we rename all of them to use
our externally genarated secret variables, and remove the
variable definitions

We will migrate the upstream secrets in the future.