Add opsfile to inject terraform variables
We use terraform to create a lot of AWS resources, and inject the
values of these resources as terraform_*.

All these variables match with variables defined in cf-deployment,
so we can replace them.

In the future we will use the naming from upstream cf-deployment
and define the variables explicitelly when calling interpolate.