Add opsfile for CC properties
This covers api, cc-worker and scheduler CC job.

We add one file per feature, and mark some changes to review
with a TODO.

Several properties in the CC jobs are set to the default value of
the spec of the job. We group them in one file for the migration
of the manifest but can be removed later.

Add CC vm-extensions for api and cc-worker