Add Datadog monitors for DNS resolution on VMs
This adds 2 monitors around DNS resolution on each VM: 1 checks that DNS
resolution works, the 2nd checks that it resolves in a timely manner.

These are being added because we had a recent issue that we eventally
tracked down to DNS resolution on one VM being flakey. In that instance
restarting the VM fixed it. Having this check will enable us to see this
sort of cause much more quickly.