Check we do not downgrade the release versions
This adds a test to ensure our rendered manifest does not contain
release versions which are lower than those specified in cf-deployment.
Now we are using a large number of ops-files it would be easy to forget
to undo the pinning of a release version - which we often do when
mitigating CVEs.