Add a syslog drain monitor
This monitors the number of syslog drains to ensure we have scaled the
adapter VMs sufficiently. According to the documentation[1] we should
scale to 3 VMs when we hit 250 drains. This is because the adapters now
cap the number of streams it runs (it runs 2 streams per drain for high
availability).

[1] https://github.com/cloudfoundry/cf-syslog-drain-release/tree/v6.4#syslog-adapter