Configure logsearch parse to consume both queues
We were configuring two different parsers consuming messages from
the specific redis queue of each AZ.

But there is no really a reason for this, and having this setup
can cause issues, as only one queue might start receiving messages
when the ingestors are redeployed and marked as unhealthy in the
ELBs. When this happens, one of the parsers will be really busy
and unable to cope with the backlog meanwhile the other is idle.

If the two parsers consume both queues, the load would be evenly
distributed across both of them and they will not be saturated.

Analysing the parser template[1], we found out that we might able
to define multiple redis "inputs" and override specific options in
each, specifically `host`. We still need to populate the property
`redis.host` with a dummy value.

With this change we do not need to configure each parser
differently and we can merge the two instance groups `parser_z1`
and `parser_z2` in one `parser`.

Additionally, we remove the static IPs in the parsers, as they are
not necessary.

[1] https://github.com/cloudfoundry-community/logsearch-boshrelease/blob/v205.0.0/jobs/parser/templates/config/input_and_output.conf.erb#L3-L21