Change paas-admin UAA_URL to login.<system>
In our deployments we were setting UAA_URL as
https://uaa<system>, which caused the IDP federation fail, as the
oauth2 clients were configured to work on the https://login.<system>
domain.

Because that, we were getting the following error when
authenticating with google:

                   Error: redirect_uri_mismatch

                   The redirect URI in the request, https://uaa.cloud.service.gov.uk/login/callback/google, does not match the ones authorized for the OAuth client. To update the authorized redirect URIs, visit: https://console.developers.google.com/apis/credentials/oauthclient/.......apps.google

The PR https://github.com/alphagov/paas-admin/pull/44 likely
fixes the issue, but changing the UAA_URL would also fix it
until that gets merged.