Remove override of cc db_logging_level
This was changed from the default of 'debug2' in 62c0882 at the same
time as changing the general logging level. This property only has any
effect if cc.log_db_queries is set to true. It defaults to false in the
boshrelease[1], and isn't being enabled anywhere in our manifests.

[1]http://bosh.io/jobs/cloud_controller_ng?source=github.com/cloudfoundry/capi-release&version=1.54.0#p%3dcc.log_db_queries