Remove rep cell Datadog monitor
After the upgrade done in [1], the rep cell monitor is broken.

The reason is that the datadog rep check does not work because uses mutual TLS. [2]

The monitor was disabled in [3] while we decided what to do about it.

We've now decided that we should remove this, as we don't believe this offers much value above the existing process monitor that's in place.

[1] https://www.pivotaltracker.com/story/show/156299675
[2] https://github.com/alphagov/paas-datadog-for-cloudfoundry-boshrelease/blob/8d8ae64cba528dc31d432f45699b3de8e100a92b/jobs/datadog-rep/templates/http_check.yaml.erb#L4-L7
[3] https://github.com/alphagov/paas-cf/pull/1317