Configure metron_agent as a manifest addon.
This removes our overrides to the base manifest.

It's been necessary to add a bunch of exclusions to the addon to prevent
it being added to all the addional VMs we add to the deployment. This
suggests that a potential future refactoring would be to separate some
of these out into separate bosh deployments.