Specify os for overridden stemcells
The base cf-deployment manifest specifies an os property for the
stemcells. The metron_agent addon uses this property to select the VMs
to apply the addon to. Because we weren't specifying an os in our
overrides, the metron agent wasn't being included anywhere.

It's not possible to specify both a name and an os for a stemcell (bosh
errors if you do), so we've had to modify the stemcell download script
to build the URL using the os parameter instead.