Remove 3rd instance of adapter and scheduler VMs.
This brings them in line with the base manifest. These were originally
created with 3 instances in 7015480, but there doesn't seem to be a
reason why.

Looking at CPU usage on these VMs over the past week in prod, the
adapter VMs peak at around 2%. The scheduler has one instance that's
consistently at around 10%, adn the others are peaking at around 3%.

Based on all this, I don't think there's a need for 3 instances of
these, so it should be safe to follow the cf-deployment config for
these.