Remove some doppler overrides.
The disable_announce and disable_syslog_drains properties were removed
in bc65bb0 because they're no longer present in the spec of the doppler
release. These properties are present in the cf-deployment manifest
though. Given they have no effect there's no need to us to remove them
with an operation (this was originally introduced to achieve a zero diff
in the manifests when switching to cf-deployment's manifest).

The shared parameter on the link allows a link to be consumed by other
bosh deployments[1]. We have no specific reason to disable this for the
doppler job.

[1]http://bosh.io/docs/links/#cross-deployment