Remove unused consul and rep cert alt names.
These are not present in the cf-deployment manifest.

The consul alt names were added in f5088bf, way back in early 2016. It's
not clear why both were added at the time, but given they're not in the
upstream manifest, and that everything works without, it makes sense to
remove them.

The rep alt name was added in c39269d, but it's not clear from the
commit that it was actually necessary. From discussions with the people
involved, it seems that several things were tried to fix an issue, but
we're not clear which one fixed it.