Stop overriding adapter_tls cn (part 1)
We want to stop overriding the common_name for this cert so that we can
use the values from upstream.

This is the first part of this change which adds the overridden CN to
the cert as an alternative_name alongside the upstream one. Once this
has been rolled out, we can do a follow-up change to remove all these
overrides. We do it this way so that this can be rolled out without
downtime.