Remove override of garden persistent_image_list
cf-depoyment sets this to the cflinuxfs2 rootfs. We added this override
because the value wan't present in our manifests when we switched to
consuming cf-deployment, and we didn't want to change anything at that
time.

There doesn't seem to be a good reason to override this, so I'm removing
it.