Stop overriding adapter_tls cn (part 2)
We want to stop overriding the common_name for this cert so that we can
use the values from upstream.

This is the second part that removes all the overrides. The first part
updated the cert to support both common_names. Now that that's rolled
out, we can remove the overrides, and use the common_name from the
cf-deployment manifest.