Deploys paas-accounts API and associated db
paas-accounts is a small API service for storing and querying details
about our users/accounts that can not be suitably stored elsewhere.

We will be initially using this to store details on who has agreed to
which terms of use documents and at what time they have agreed. The API
will be consumed privately by paas-admin protected behind basic auth.