Remove overrides of webdav blobstore config.
We use the upstream `use-s3-blobstore.yml` opsfile that removes the
blobstore_type property, meaning the default value of "fog" is used. As
a result all the webdav properties don't apply, so there's no need to
make changes to these settings.

These overrides were originally added to achieve a zero-diff deploy when
we initially migrated to the cf-deployment manifests. Retaining these
overrides only adds to the changes we need to maintain.