Remove cloudfoundry group from pipeline
This was identical to the 'All' group except for the pipeline-lock job
which isn't included. It therefore doesn't offer much value, so let's
remove it.

This also renames the 'all' group to 'depoly' because it doesn't contain
all jobs, so the name is misleading.

This is leftover from when this pipeline used to deploy both bosh and
cloudfoundry.