healthcheck/es: Content-Type request header
This appears to be required for newer Elasticsearch versions provided by
Aiven (which we're spiking against):

    {"error":"Content-Type header [] is not supported","status":406}

Adding it won't have any negative effect for older versions. Strictly
it may not be required for all requests, but it's correct for the
current set of requests and shouldn't cause harm in the future.