healthcheck: Allow CA cert to be empty
Return a default `tls.Config` when no CA cert is provided.

We're testing an Elasticsearch service from Aiven which uses publicly
trusted certs for service instances so we don't need to populate
`CACertBase64` in the bind credentials.

I considered changing the way that the client was constructed, to only
set `client.client.Transport.TLSClientConfig` when the CA is a non-empty
string but the code looked much more complicated than modifying
`buildTLSConfigWithCACert` and it wouldn't have applied to MongoDB which
also uses this function.