Import datadog addon from runtime config.
We're moving away from having addons defined in the global
runtime-config because this has to be defined in a separate repo which
means that changes to it have to be manually coordinated and deployed
across the 2 repos.

This imports the datadog addon that's currently configured there.

I've also added caching of datadog manifest in the tests which speeds
them up considerably.