Avoid duplication in generate-cloud-config task
We were devining all these values as PARAMS that were then just
interpolated into the file. Meaning that new values have to be added in
both places. It's much simpler to inline the interpolations in the task
body.

It's been necessary to quote the values in the pipeline upload script so
that when they're not set they are passed through as empty string
instead of nil. Without this change, concourse returned the following
error:

could not resolve template vars: Invalid type '<nil>' for value '<nil>'
and variable 'oauth_client_id'. Supported types for interpolation within
a string are integers and strings.