Use consul instead of bosh-dns for links
log-cache release relies on bosh-dns[1] to disvoer where the
reverse_log_proxy is located, using a `.bosh` hostname. But we
do not use BOSH DNS yet.

To work around this, we override the link information using a consul
service domain for some services like reverse_log_proxy and
cloud_controller.

We tried using `ip_addresses: true`[2] but it had no effect, possibly
because the instance group is on a dynamic network.

You have to specify `instances` when specifying `properties`, but it can
be an empty slice because the release uses `address` instead
(https://bosh.io/docs/links-manual/)

Consul agent needs to be run on the same instance group in order to
resolve the internal service hostname that RLP publishes. We assume
that is the case for doppler.

Original commit: https://github.com/alphagov/paas-cf/commit/785aa3c6be10a1548dba1d569821bf99a3dcc548

[1] https://bosh.io/docs/dns
[2] https://bosh.io/docs/dns/#links