Don't disable the uaa.require_https flag.
The default is true upstream but we've set it to false in [1].

If require_https is disabled the X-Uaa-Csrf cookie won't have the secure flag
which came up in our latest pentest.

There is no reason to disable this anymore. Whatever caused issues previously
it doesn't seem to be present anymore.

[1] https://github.com/alphagov/paas-cf/commit/9a4dda67ead70bd39d811cc373d8fd10dbb9df44