Disable all unencrypted RDS plans
We want to deprecated the old unencrypted RDS plans, so we will
explicitly disable them.

We are explicit when enabling RDS plans because
using just `cf enable-service-access <service>` enables all
the backing services in that broker, and later we disable the ones
that are deprecated. If we do it that way, the user will see the
deprecated services for some seconds until they are disabled
in every deployment.

Instead, we enable the service plans one by one on demand.

We also test that only the supported plans are available for the users.
To run this test, we must be sure that we run cf marketplace as a
normal user.