Check if healthcheck-db exists before creating
When creating the healthcheck-db We were relying on the behaviour that
`cf create-service` is a no-op when creating a plan with the exact same
name, plan and args.

However we recently renamed the plan which causes the create-service
call to fail.

To resolve this a guard is added to check for the existance of the
database before attempting to create it.