Extend rotate-cf-certs script to regenerate CA certs with SKIs
BOSH cli can be used to generate certificates into a variable store - based on
the manifest variable definitons.
There is a known issue if you try to rotate the CA certificates outlined in [1].

The root of the problem is that when OpenSSL is configured to trust multiple
CAs, and two of them have the same subject name, OpenSSL will only verify
certificates against the first one (see [2] in OpenSSL code).

One solution for the CA certificate rotation problem is to set the Subject Key
Identifiers so OpenSSL will be able to handle multiple certificates with the
same subject name.

We patched the bosh-cli to generate certificates with an SKI. This will solve
the problem long-term but the first rotation will still cause downtime as
OpenSSL is still not able to find the right CA unless all CAs with the same
common name have SKIs.

To achieve the first rotation without downtime (when old CAs don't have SKIs
but new CAs have it) we want to regenerate the CA certificates using the
original private keys so we can add SKIs.

[1] https://docs.google.com/document/d/1vKxziTEvIgKHubukoyAGaJzGqrMBjun7JffbunlLBPg/edit#heading=h.wftyivqbaag4
[2] https://github.com/openssl/openssl/blob/49f6cb968ff63793f6671d9026fb2a7034dad79a/crypto/x509/x509_lu.c#L613-L617