Don't stop reporting domain metrics when any of the domain TLS check fails
When any of the domain TLS check has an error we just log it and continue with the next one.
Also we changed the DaysUntilExpiry method so it always returns an error when a
certificate has a non-expiry related issues.

We only want to be alerted when correctly set up domains are near expiry.

When tenants don't set up DNS for their domain, or point it to the wrong domain
we shouldn't get an alert.