Unpin our fork of cf-smoke-tests
The original reason for maintaining a fork has now been merged[1].

We should be using the version of the smoke tests referenced in
cf-deployment, which pins the release repository[2], not the
smoke test repository itself. Therefore, our Concourse pipeline
is clearer if we pull in the release repository with the matching
version and use the smoke test submodule.

[1] https://github.com/cloudfoundry/cf-smoke-tests/pull/39
[2] https://github.com/cloudfoundry/cf-smoke-tests-release