Fix log-cache CA cert rotation.
Because some of these entries were referencing the certs via the
`certificate.ca` variable instead of the `ca_certificate.certificate`
variable, they weren't picked up by the test we have to detect this.
As a result, they were in use without the old CA cert also deployed.
This led to downtime while rotating the CA certs.

This extends the test to catch this indirect reference to the CA cert,
and fixes the rotation for these certs.

As part of this, I've moved the rotation ops entries into
900-cf-cert-rotation to keep them with all the other ones.