Version pin paas-admin
We're implementing the version bumping for that application from the
build pipeline. We should fully make use of it in order to avoid any
incompatibilities in the future and run the desired version of the
application across all of our environments.

This does however introduces the inconvenience of the need to come into
this repo, and manually bump the version as we go.