Fix pinning of paas-cf for continuous smoke tests
The continuous-smoke-tests job relies on a task defined in the paas-cf
repo. Because of this it's necessary for the pipeline and the repo to
match, otherwise there will be problems like the task expecting
different input names to those provided by the pipeline.

When the smoke tests were changes to pin the release repo, the input
name was changed from `cf-smoke-tests` to `cf-smoke-tests-release`. With
the pinning as it stands, there was a period between the
self-update-pipelines task, and the post-deploy completing where the
continuous smoke tests were failing with the error "missing inputs:
cf-smoke-tests" due to the mismatch.

Updating the pinning to the pipeline-lock job means that it will use the
version of paas-cf that's just be used to run self-update-pipelines, and
therefore won't hit these issues.