Be more precise in matching paas-cf tags.
With the new Longon deployments, we're going to have the tag_prefix in
some environments be the prefix of ones in other environments (eg prod
and prod-lon). We don't want tags from prod-lon to be matched by the
prod deployment. Additionally, we don't want the tag_release script to
see existing tags from one environment, and therefore skip doing that
tagging because it thinks the repo is already tagged.