Refactor how we pass around tag prefixes
We were specifying a tag prefix, and a filter in the Makefile, but it
wasn't easy to see what these controlled. Renaming these to an input and
output prefix should make it clearer that one is used to determine which
tags to use as input, and the other controls the tags created as an
output from the pipeline.

Changing from specifying a filter to a prefix for the input is done to
prepare for the next change to allow us to be more specific in how we
match input tags.