Fix healthcheck DB creation script.
The format of the output from `cf service` has changed with a recent CF
upgrade. The status line is now:

    status:    create succeeded

This meant that when creating the database, the script was stalling
forever waiting for the status to to match. To make this more robust we
now simply check for 'create succeeded' anywhere in the output.