Fix Pingdom contact ids
The ids for the contacts in Pingdom appear to have changed since we last
ran this terraform. The existing checks are still linked to the correct
contacts, it's just that the internal IDs of these are now different.
This means that future terraform runs will attempt to set these back to
the old ids, which will fail because they no longer exist.