Remove `tps_watcher` "locks held" monitor
After migrating to Locket `tps_watcher` stopped reporting the number of
locks it is holding and it doesn't appear to be providing an equivalent
to migrate to.

After discussion with @alext and @keymon we've decided to remove the
monitor as we're not sure it's monitoring something that absolutely must
be monitored.

`tps_watcher` should probably be using the `LockHeldMetronNotifier`[1]
in Locket.

[1] https://github.com/cloudfoundry/locket/blob/052dc91cec4c80cd1c888638fd831ce1f990a970/lockheldmetrics/lock_held_metron_notifier.go#L25