Split out log-cache acceptance tests
We run the "apps" suite of the acceptance tests enabling using
cf tail to test log-cache[1][2]

But this feature is quite new and in development and the tests
might be flaky. [3]

We will separate the execution of these tests to a different job
so they can be rerun independently from the more lenghtly acceptance
tests.

We remove the artifact directory, as we do not upload the artifacts.

[1] https://github.com/alphagov/paas-cf/pull/1371
[2] https://github.com/alphagov/paas-cf/pull/1371/commits/d141e72dc214f3cbda02855014faa6d66438c532
[3] https://github.com/alphagov/paas-cf/pull/1410