desensitizes doppler elb latency monitor
We are consistantly seeing our Doppler ELB Latency monitor triggered as
a result of the anomaly detection algorithm noticing large %-changes of
what are ultimately very small values.

This alters the monitor configuration to use the 'agile' detection
algorithm (which takes into account a larger history of data), uses a
smaller window for sampling and uses warning/alert thresholds at a
25%-change and 50%-change respectfully.

The result should be a monitor that is less sentitive to short spikes of
%-change but still trigger on more sustained deviations.

This may mean that it takes a bit longer for an alert to trigger and may
potentially be prone to miscalculation if the value rises very slowly over
say a few weeks. However we have not seen such behaviour and that could
be caught using a more traditional "alert if value > x" style monitor.