Increase rds-broker instance type to small
We have added the rds-metrics-collector job in the rds-broker
instance-group to collect tenants metrics. But this job runs
periodically (every 1-5 minutes) instead of in a burst mode as
the broker use to. Because that the VM ends running out of credits.

We will try to increase with the `small` size, which is a
t2.small, and see if it behaves better with it.