Add scripts to rename/rename cf-secrets to cf-vars
We want to start using bosh to generate the passwords used in the
bosh manifest, so that we are more aligned with the upstream
cf-deployment.

The passwords generated by bosh with bosh interpolate will be
stored in the cf-vars-store.yml file in S3, which is a plain
`key: value` similar to cf-secrets.yml

We want to move here as many secrets as possible to cf-vars-store.yml
but there are some cases we need to keep in cf-secrets:

   - vcap user password for cloud config, that needs
     to be passed as a SHA512 encoded password.
     We can actually stop setting this password, as the
     base stemcells set a random password by default
     as in [1]

   - The RDS passwords: The master password is used by terraform
     and then each password is configured in the create-cf-dbs.sh
     script. cf-deployment expects these variables to be passed
     as `external_*`[2], etc.

We also want to rename variables to align them to the upstream
names used in the original cf-deployment.

To do this migration, we add two scripts and a task that calls
them before the generate-cf-secrets:

 - migrate-cf-secrets.rb: will move and rename a given
   set of variables from cf-secrets.yml to cf-vars-store.yml.
   It also does some conversion for special cases with ssh
   keys, lists of passwords, or one-to-many mapping.
 - rename-cf-secrets.yml: renames variables in cf-secrets.yml

The task contains the list of all the variables migrated and/or
renamed and their new names.

We keep in generate-cf-secrets.rb the secrets that we want to
perserve in cf-secrets.yml, otherwise the script deletes them.
There are a part of secrets that are not needed anymore:

 - secrets_loggregator_endpoint_shared_secret
 - secrets_uaa_cc_client_secret

These scripts and task would be used for the migration, and removed later.

[1] https://www.pivotaltracker.com/story/show/114110115
[2] https://github.com/cloudfoundry/cf-deployment/blob/72c064181ab9420126b248cf43d6f0f97ced6394/operations/use-external-dbs.yml

Signed-off-by: henrytk <henry.knott@digital.cabinet-office.gov.uk>
Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>