Remove dedictated uaa_jwt_signing_key rotation
We would use the rotate-vars-store-secrets.rb script to rotate
both the uaa_jwt_signing_key and the uaa_jwt_signing_key_id

We need to also rename the variable uaa_jwt_signing_key_id_old,
so that the name has a _old prefix and would be rotated with
the secrets when using the rotate-vars-store-secrets.rb script.

Follow 67c6238cc84712abf07679b666ddcc2b2c02920f for more info
about the rotation of this script.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>