Rotate vars-store script to support all types
We want to be able to rotate any type of credentials store in
cf-vars-store, not only certificates.

We also want to support the same logic for the _old secrets:
if there is a _old secret defined then rename the variable,
otherwise delete it.

We do a big refactoring of the tests and implementation.

Signed-off-by: henrytk <henry.knott@digital.cabinet-office.gov.uk>