Do not rotate secrets that were created by bosh
Now we moved the generation and rotation of secrets to
bosh-interpolate + cf-vars-store.yml, we are rotating passwords
that were previously generated by bosh interpolate but not
rotated.

We need to study each case to find out if they are OK to rotate
without downtime.

For the time being we limit to not rotate them, and we will
revisit this in a follow up story.