Upgrade rds-broker release to 0.1.36
Includes the changes from
 - https://github.com/alphagov/paas-rds-broker-boshrelease/pull/61
 - https://github.com/alphagov/paas-rds-metric-collector/pull/5
 - https://github.com/alphagov/paas-rds-broker/pull/83

We are experiencing rate limiting throttling in our RDS brokers,
very likely due the amount of calls to the ListTagsForResource
operation.

This version would cache the response and efectivelly reduce
the amount of requests to the AWS api.

Also includes change the logging level for the metrics collector
as in:
 - https://github.com/alphagov/paas-rds-broker-boshrelease/pull/62