Change broker cronjobs to run daily.
These presently, only check for final snapshots older than 35 days to be
deleted. As such it doesn't make sense to run this job hourly - daily is
sufficient.

While it may appear at first glance that the existing configuration was
running this daily, the broker is actually using a cron library[1] which
has second-level precision, and therefore has a 6 field syntax (with the
last one, day-of-week, being optional), meaning that the existing config
was specifying 0 for the seconds and minutes, and would therefore run
every hour.

[1]https://godoc.org/github.com/robfig/cron