Pin a specific version of the Alpine containers.
The latest tag was recently updated to point to the 3.8 release, but
this release was broken[1]. attempting to pull the container locally
resulted in the error:

$ docker pull alpine:latest
latest: Pulling from library/alpine
no matching manifest for linux/amd64 in the manifest list entries

This has caused the continuous smoke tests to fail to start in all
environments as of around 9:40 today.  Pinning this to the version we
were previously using (3.7) should help avoid this sort of issue in
future.

[1]https://github.com/gliderlabs/docker-alpine/issues/418