Add monitors for smoke test errors and no data
These are just for the continuous smoke tests. We had an issue where
they were erroring, but the existing monitor only checks for failure
states.

These have to be done as two separate checks because of the way Datadog
handles monitors. If you add a condition to alert on no data, then the
absence of metrics with the `build_status:errored` tag will cause the
monitor to persistently stay triggered.

For the no data monitor, we have to assert on some query to be able to
create the monitor, but we really only care about the no data condition.
However, when the monitor is receiving no data the evaluation window
cannot be calculated, and therefore the alert condition does _not_
trigger the monitor after 15 minutes, as may be expected.