Use Bosh to generate Aiven broker password
This will be used by Cloud Foundry for basic authentication with the
broker. The broker will be deployed as an app on the platform. This
means nothing in the manifest consumes this secret. It will be extracted
from the manifest in the pipeline.

This secret will not be rotated. I have done this to be consistent with
what we have done with the Compose broker. However, we may want to
reconsider this in future, because there is only a small amount of time
between deploying the broker and updating its registration with Cloud
Foundry, so the downtime would be very small.