Rename Compose Elasticsearch service
We want this name for the Aiven Elasticsearch service. The registration
of the Compose broker has been moved into the deploy task because it
doesn't need to be in a separate task, and this means the service gets
renamed before the Aiven broker tries to use the same name.

For the same reason we have moved the deployment of the Aiven broker
into the next aggregate block, otherwise it will run in parallel with
the deployment of the Compose broker.