Custom acceptance tests: parse ES URLs properly
Using ad hoc string manipulation on URLs strikes again! This was causing
test failure for Aiven tests because it assumes there will be a forward
slash at the end of the host(:port) part of the URL.

Using a well-tested library is much preferred as it takes away the
effort and pain of caring about these little details.